compuation work sepearte from app componet to a hook
